Relocation of DiagramId, Filter, Response, AnalyticalProperty, and TwoBodyChannel to FrontEnds Module.
