+++ src/frontend/GV.jl
@@ -6,31 +6,23 @@ import ..ComputationalGraphs: FeynmanGraph
 import ..ComputationalGraphs: Graph
 import ..ComputationalGraphs: _dtype
 import ..Parquet
-import ..Parquet: Filter, NoBubble, NoHartree, NoFock, DirectOnly
-import ..Parquet: Wirreducible  #remove all polarization subdiagrams
-import ..Parquet: Girreducible  #remove all self-energy inseration
-import ..Parquet: NoBubble  # true to remove all bubble subdiagram
-import ..Parquet: Proper  #ver4, ver3, and polarization diagrams may require to be irreducible along the transfer momentum/frequency
-import ..Parquet: Response, Composite, ChargeCharge, SpinSpin, UpUp, UpDown
-import ..Parquet: AnalyticProperty, Instant, Dynamic, D_Instant, D_Dynamic
 import ..Parquet: DiagramType, VacuumDiag, SigmaDiag, GreenDiag, PolarDiag, Ver3Diag, Ver4Diag
+import ..Parquet: Interaction, DiagPara
 import ..Taylor
 using ..FrontEnds
+import ..FrontEnds: Filter, NoHartree, NoFock, DirectOnly
+import ..FrontEnds: Wirreducible  #remove all polarization subdiagrams
+import ..FrontEnds: Girreducible  #remove all self-energy inseration
+import ..FrontEnds: NoBubble  # true to remove all bubble subdiagram
+import ..FrontEnds: Proper  #ver4, ver3, and polarization diagrams may require to be irreducible along the transfer momentum/frequency
+import ..FrontEnds: Response, Composite, ChargeCharge, SpinSpin, UpUp, UpDown
+import ..FrontEnds: AnalyticProperty, Instant, Dynamic
+import ..FrontEnds: DiagramId, Ver4Id, Ver3Id, GreenId, SigmaId, PolarId, BareGreenId, BareInteractionId
+
+
 using AbstractTrees
 
 import ..Utility: taylorexpansion!
-
-import ..Interaction
-import ..DiagPara
-
-import ..DiagramId
-import ..Ver4Id
-import ..Ver3Id
-import ..GreenId
-import ..SigmaId
-import ..PolarId
-import ..BareInteractionId
-import ..BareGreenId
 
 export eachorder_diag, diagdictGV, diagdict_parquet, leafstates, leafstates_diagtree
