+++ test/taylor.jl
@@ -1,6 +1,8 @@
 using FeynmanDiagram
 using FeynmanDiagram: Taylor as Taylor
 using FeynmanDiagram: ComputationalGraphs as Graphs
+import FeynmanDiagram.Parquet: DiagPara
+import FeynmanDiagram.FrontEnds: DiagramId, PropagatorId, GenericId, Ver4Id, Ver3Id, GreenId, SigmaId, PolarId, BareGreenId, BareInteractionId
 
 function assign_random_numbers(g, taylormap1, taylormap2) #Benchmark taylor expansion generated by two different methods
     leafmap1 = Dict{Int,Int}()
@@ -196,7 +198,7 @@ end
     # autodiff
     factor = 1 / (2π)^D
     set_variables("x y"; orders=[2, 2])
-    propagator_var = Dict(BareGreenId => [true, false], BareInteractionId => [false, true]) # Specify variable dependence of fermi (first element) and bose (second element) particles.
+    propagator_var = Dict(BareGreenId{DiagPara} => [true, false], BareInteractionId{DiagPara} => [false, true]) # Specify variable dependence of fermi (first element) and bose (second element) particles.
     t, taylormap = taylorexpansion!(root, propagator_var)
 
     taylorleafmap, taylorleafvec = assign_leaves(root, taylormap)
